requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
beautifulsoup4==4.11.2
click==8.1.3
coverage==7.2.6
decorator==5.1.1
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==3.2.18
django-bootstrap-v5==1.0.11
django-cache-url==3.4.4
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==22.1
django-redis==5.2.0
django-rq==2.7.0
djangorestframework==3.14.0
environs==9.5.0
executing==1.2.0
factory-boy==3.2.1
Faker==17.0.0
fontawesomefree==6.3.0
gunicorn==20.1.0
ipython==8.13.0
jedi==0.18.2
Markdown==3.4.1
marshmallow==3.19.0
matplotlib-inline==0.1.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
rq==1.13.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
uWSGI==2.0.21
wcwidth==0.2.6
whitenoise==6.2.0
